requirements.txt
openai
dotenv
pillow
pyautogui
python-docx
